+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [3.7.1](https://github.com/Backbase/stream-services/compare/3.6.0...3.7.1)
 ### Added
 - Add **portfolio-http**
-- Add `POST /integration-api/v1/regions/batch` endpoint to portfolio
+- Add `POST /portfolios/regions/batch` endpoint to portfolio
 
 ## [3.6.0](https://github.com/Backbase/stream-services/compare/3.5.0...3.6.0)
 ### Added